+++ evm/src/main/java/com/horizen/evm/library/AccountParams.java
@@ -1,0 +1,13 @@
+package com.horizen.evm.library;
+
+public class AccountParams extends HandleParams {
+    public String address;
+
+    public AccountParams() {
+    }
+
+    public AccountParams(int handle, String address) {
+        super(handle);
+        this.address = address;
+    }
+}
