+++ evm/src/main/java/com/horizen/evm/library/EvmResult.java
@@ -1,0 +1,8 @@
+package com.horizen.evm.library;
+
+public class EvmResult {
+    public byte[] returnData;
+    public String address;
+    public long leftOverGas;
+    public String evmError;
+}
