+++ evm/src/main/java/com/horizen/evm/library/HandleParams.java
@@ -1,0 +1,12 @@
+package com.horizen.evm.library;
+
+public class HandleParams extends JsonPointer {
+    public int handle;
+
+    public HandleParams() {
+    }
+
+    public HandleParams(int handle) {
+        this.handle = handle;
+    }
+}
