+++ evm/src/main/java/com/horizen/evm/library/InitializeParams.java
@@ -1,0 +1,12 @@
+package com.horizen.evm.library;
+
+public class InitializeParams extends JsonPointer {
+    public String path;
+
+    public InitializeParams() {
+    }
+
+    public InitializeParams(String path) {
+        this.path = path;
+    }
+}
