+++ evm/src/main/java/com/horizen/evm/library/NonceParams.java
@@ -1,0 +1,13 @@
+package com.horizen.evm.library;
+
+public class NonceParams extends AccountParams {
+    public long nonce;
+
+    public NonceParams() {
+    }
+
+    public NonceParams(int handle, String address, long nonce) {
+        super(handle, address);
+        this.nonce = nonce;
+    }
+}
