+++ evm/src/main/java/com/horizen/evm/library/OpenStateParams.java
@@ -1,0 +1,12 @@
+package com.horizen.evm.library;
+
+public class OpenStateParams extends JsonPointer {
+    public String root;
+
+    public OpenStateParams() {
+    }
+
+    public OpenStateParams(String root) {
+        this.root = root;
+    }
+}
